requirements.txt
Pillow==9.0.1
Telethon==1.24.0
python-socks==2.0.3
asyncio==3.4.3
typing==3.7.4.3